pyproject.toml
[project]
name = "lx-music-api-server"
version = "3.0.2"
description = "LX Music Api Server"
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.119.0)",
    "pycryptodome (>=3.22.0,<4.0.0)",
    "xmltodict (>=0.14.2,<0.15.0)",
    "redis (>=5.2.1,<6.0.0)",
    "cryptography (>=44.0.3,<47.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "dnspython (>=2.8.0,<3.0.0)",
    "aiohttp (>=3.12.15,<4.0.0)",
    "loguru>=0.7.3",
    "orjson>=3.11.3",
]
